finished disk communication

See merge request matanKabalo/matan-yonatan-os!33

+++ src/kernel/drivers/screen.cpp
@@ -116,6 +116,7 @@ void ScreenApi::clearLastPixels(int rows, int amount) {
             drawPixel(x, y, COLOR_BLACK);
             y++;
         }
+        
 
         y = getCursur().y;
         x--;
@@ -401,7 +402,7 @@ void ScreenApi::clearCursor()
     _cursorShow = false;
 }
 
-const uint64_t ScreenApi::getFbStartAddress()
+const uint32_t ScreenApi::getFbStartAddress()
 {
     return _fbInfo->addr;
 }
